🚚 Moves / organises service utils into sub-directories
