+++ services/endpoints/ping-check.js
