+++ tests/components/item.test.js
@@ -5,7 +5,7 @@ import { shallowMount } from '@vue/test-utils';
 import { createStore } from 'vuex';
 import Item from '@/components/LinkItems/Item.vue';
 import router from '@/router';
-import pingCheck from '../../services/ping-check';
+import pingCheck from '../../services/endpoints/ping-check';
 
 vi.mock('@/utils/request', () => {
   const fn = vi.fn(() => Promise.resolve({ data: {} }));
