+++ tests/server/conf-strip.test.js
@@ -35,7 +35,7 @@ beforeAll(() => {
   fs.writeFileSync(path.join(tmpDir, 'sub.yaml'), 'pageInfo:\n  title: Sub\n');
   process.env.USER_DATA_DIR = tmpDir;
   delete require.cache[require.resolve('../../services/app')];
-  delete require.cache[require.resolve('../../services/auth-oidc')];
+  delete require.cache[require.resolve('../../services/utils/auth-oidc')];
   app = require('../../services/app');
 });
